# src/api/rest-api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorModel": {
        "properties": {
          "error": {
            "description": "Error type",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Detailed error message",
            "example": "Validation failed",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code",
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "ProjectModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the project was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique project identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "name": {
            "description": "URL-friendly project name (lowercase letters, numbers, hyphens)",
            "example": "my-blog",
            "type": "string"
          },
          "organizationId": {
            "description": "Organization that owns this project",
            "example": "acme-corp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organizationId",
          "createdAt",
          "name"
        ],
        "type": "object"
      },
...
      "RevisionModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the revision was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique revision identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "isDraft": {
            "description": "Whether this is the draft revision. Draft revisions can be modified; use for all write operations.",
            "type": "boolean"
          },
          "isHead": {
            "description": "Whether this is the head (latest committed) revision. Head is immutable and represents the published state.",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "createdAt",
          "isDraft",
          "isHead"
        ],
        "type": "object"
      },
...
      "SuccessModelDto": {
        "properties": {
          "success": {
            "example": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "UpdateProjectDto": {
        "properties": {
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "isPublic"
        ],
        "type": "object"
      },
...
      "UpdateTableDto": {
        "properties": {
          "patches": {
            "items": {
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "patches"
        ],
        "type": "object"
      },
      "UpdateTableResponse": {
        "properties": {
          "previousVersionTableId": {
            "type": "string"
          },
          "table": {
            "$ref": "#/components/schemas/TableModel"
          }
        },
        "required": [
          "previousVersionTableId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Revisium API",
    "version": "2.6.0-alpha.3"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organization/{organizationId}/projects/{projectName}": {
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Project"
        ]
      },
      "get": {
        "operationId": "project",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get project by name",
        "tags": [
          "Project"
        ]
      },
      "put": {
        "operationId": "updateProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update project settings",
        "tags": [
          "Project"
        ]
      }
    },
...
    "/api/revision/{revisionId}": {
      "get": {
        "operationId": "revision",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get revision by ID",
        "tags": [
          "Revision"
        ]
      }
    },
...
    "/api/revision/{revisionId}/tables/{tableId}": {
      "delete": {
        "operationId": "deleteTable",
        "parameters": [
          {
            "description": "Draft revision identifier (modifications are only allowed in draft revisions)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly table identifier",
            "in": "path",
            "name": "tableId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Table not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a table",
        "tags": [
          "Table"
        ]
      },
      "get": {
        "operationId": "table",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly table identifier",
            "in": "path",
            "name": "tableId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Table not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get table by ID",
        "tags": [
          "Table"
        ]
      },
      "patch": {
        "operationId": "updateTable",
        "parameters": [
          {
            "description": "Draft revision identifier (modifications are only allowed in draft revisions)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly table identifier",
            "in": "path",
            "name": "tableId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTableDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTableResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Table not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update table schema using JSON Patch",
        "tags": [
          "Table"
        ]
      }
    },
...
  },
  "servers": [],
...
}